refactor: Unified asset tag template form for create and edit
- Created single AssetTagTemplateForm component with optional templateId prop
- Removed separate create and edit form components (588 lines eliminated)
- Applied improved 3-column layout from create form to both modes:
  * Left column (8/12): Organized cards for input sections
  * Right column (4/12): Sticky live preview + template codes reference
- Edit mode auto-loads existing template data via useEffect
- Both pages (new + [id]/edit) now use same component with consistent UX
- Updated page layouts to match with breadcrumb navigation
- TypeScript compilation passes
- ESLint passes
