--- components/forms/asset-tag-template-edit-form.tsx
@@ -1,345 +1,0 @@
-
-'use client';
-
-import { useForm, useFieldArray } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import * as z from 'zod';
-import { AssetTagTemplatePreview } from '@/components/asset-tag-templates/template-preview';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { createClient } from '@/lib/supabase/client';
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { AssetTagTemplate, AssetTagTemplateElement } from '@/components/asset-tag-templates/types';
-import { useEffect } from 'react';
-
-const preprocessOptionalNumber = (min?: number) =>
-  z.preprocess((v) => {
-    if (v === '' || v === null || (typeof v === 'number' && Number.isNaN(v))) return undefined;
-    return v;
-  }, (min !== undefined ? z.number().min(min) : z.number()).optional());
-
-const formSchema = z.object({
-  name: z.string().min(1, 'Name is required'),
-  description: z.string().optional(),
-  tagWidthMm: z.number().min(1, 'Width must be at least 1'),
-  tagHeightMm: z.number().min(1, 'Height must be at least 1'),
-  marginMm: preprocessOptionalNumber(0),
-  backgroundColor: z.string().optional(),
-  textColor: z.string().optional(),
-  borderColor: z.string().optional(),
-  borderWidthMm: preprocessOptionalNumber(0),
-  textSizePt: preprocessOptionalNumber(1),
-  isMonochrome: z.boolean().optional(),
-  prefix: z.string().optional(),
-  numberLength: preprocessOptionalNumber(1),
-  suffix: z.string().optional(),
-  numberingScheme: z.enum(['sequential', 'random']).optional(),
-  stringTemplate: z.string().optional(),
-  codeType: z.enum(['QR', 'Barcode', 'None']).optional(),
-  codeSizeMm: preprocessOptionalNumber(1),
-  elements: z.array(z.object({
-    type: z.enum(['text', 'qrcode', 'barcode', 'image']),
-    x: z.number(),
-    y: z.number(),
-    value: z.string(),
-    size: preprocessOptionalNumber(1),
-    height: preprocessOptionalNumber(1),
-    color: z.string().optional(),
-  })),
-});
-
-type FormValues = z.infer<typeof formSchema>;
-
-interface AssetTagTemplateEditFormProps {
-  templateId: number;
-}
-
-export function AssetTagTemplateEditForm({ templateId }: AssetTagTemplateEditFormProps) {
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const router = useRouter();
-  const supabase = createClient();
-
-  const form = useForm({
-    resolver: zodResolver(formSchema),
-  });
-
-  useEffect(() => {
-    async function fetchTemplate() {
-      const { data, error } = await supabase
-        .from('asset_tag_templates')
-        .select('template')
-        .eq('id', templateId)
-        .single();
-      if (error || !data) {
-        console.error('Failed to fetch template data', error);
-        router.push('/management/asset-tag-templates');
-      } else {
-  form.reset(data.template as unknown as FormValues);
-      }
-    }
-    fetchTemplate();
-  }, [templateId, supabase, router, form]);
-
-  const { fields, append, remove } = useFieldArray({
-    control: form.control,
-    name: 'elements',
-  });
-
-  const onSubmit = async (raw: unknown) => {
-    const values = raw as FormValues;
-    setIsLoading(true);
-    setError(null);
-    
-    try {
-      const templatePayload: AssetTagTemplate = {
-        name: values.name,
-        description: values.description,
-        tagWidthMm: values.tagWidthMm,
-        tagHeightMm: values.tagHeightMm,
-        marginMm: values.marginMm as number | undefined,
-        backgroundColor: values.backgroundColor,
-        textColor: values.textColor,
-        borderColor: values.borderColor,
-        borderWidthMm: values.borderWidthMm as number | undefined,
-        textSizePt: values.textSizePt as number | undefined,
-        isMonochrome: values.isMonochrome,
-        prefix: values.prefix,
-        numberLength: values.numberLength as number | undefined,
-        suffix: values.suffix,
-        numberingScheme: values.numberingScheme,
-        stringTemplate: values.stringTemplate,
-        codeType: values.codeType,
-        codeSizeMm: values.codeSizeMm as number | undefined,
-        elements: values.elements.map(e => ({
-          ...e,
-          size: e.size as number | undefined,
-          height: e.height as number | undefined,
-        })),
-      };
-
-      const { error: updateError } = await supabase
-        .from('asset_tag_templates')
-        .update({ template: templatePayload })
-        .eq('id', templateId);
-
-      if (updateError) {
-        throw updateError;
-      }
-
-      console.log('Template updated successfully');
-      router.push('/management/company-settings?tab=templates');
-      
-    } catch (err) {
-      console.error('Error updating template:', err);
-      setError(err instanceof Error ? err.message : 'Failed to update template');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-      {error && (
-        <div className="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded">
-          {error}
-        </div>
-      )}
-      
-      <div className="space-y-4">
-        <h3 className="text-lg font-medium">Basic Information</h3>
-        <div>
-          <label className="block text-sm font-medium mb-1">Name</label>
-          <Input placeholder="My Template" {...form.register('name')} />
-        </div>
-        <div>
-          <label className="block text-sm font-medium mb-1">Description</label>
-          <Input placeholder="Optional description" {...form.register('description')} />
-        </div>
-      </div>
-
-      <div className="space-y-4">
-        <h3 className="text-lg font-medium">Dimensions</h3>
-        <div className="grid grid-cols-2 gap-4">
-          <div>
-            <label className="block text-sm font-medium mb-1">Width (mm)</label>
-            <Input type="number" {...form.register('tagWidthMm', { valueAsNumber: true })} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Height (mm)</label>
-            <Input type="number" {...form.register('tagHeightMm', { valueAsNumber: true })} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Margin (mm)</label>
-            <Input type="number" step="0.1" {...form.register('marginMm', { valueAsNumber: true })} />
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-4">
-        <h3 className="text-lg font-medium">Styling</h3>
-        <div className="grid grid-cols-2 gap-4">
-          <div>
-            <label className="block text-sm font-medium mb-1">Background Color</label>
-            <Input type="color" {...form.register('backgroundColor')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Text Color</label>
-            <Input type="color" {...form.register('textColor')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Border Color</label>
-            <Input type="color" {...form.register('borderColor')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Border Width (mm)</label>
-            <Input type="number" step="0.1" {...form.register('borderWidthMm', { valueAsNumber: true })} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Text Size (pt)</label>
-            <Input type="number" {...form.register('textSizePt', { valueAsNumber: true })} />
-          </div>
-          <div className="flex items-center space-x-2">
-            <input type="checkbox" {...form.register('isMonochrome')} />
-            <label className="text-sm font-medium">Monochrome (Black & White)</label>
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-4">
-        <h3 className="text-lg font-medium">Code Generation</h3>
-        <div className="grid grid-cols-2 gap-4">
-          <div>
-            <label className="block text-sm font-medium mb-1">Prefix</label>
-            <Input placeholder="e.g., EQ" {...form.register('prefix')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Number Length</label>
-            <Input type="number" {...form.register('numberLength', { valueAsNumber: true })} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Suffix</label>
-            <Input placeholder="e.g., A" {...form.register('suffix')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Numbering Scheme</label>
-            <select {...form.register('numberingScheme')} className="w-full px-3 py-2 border border-gray-300 rounded-md">
-              <option value="sequential">Sequential</option>
-              <option value="random">Random</option>
-            </select>
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">String Template</label>
-            <Input placeholder="e.g., {prefix}-{number}-{suffix}" {...form.register('stringTemplate')} />
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Code Type</label>
-            <select {...form.register('codeType')} className="w-full px-3 py-2 border border-gray-300 rounded-md">
-              <option value="QR">QR Code</option>
-              <option value="Barcode">Barcode</option>
-              <option value="None">None</option>
-            </select>
-          </div>
-          <div>
-            <label className="block text-sm font-medium mb-1">Code Size (mm)</label>
-            <Input type="number" {...form.register('codeSizeMm', { valueAsNumber: true })} />
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-4">
-        <h3 className="text-lg font-medium">Elements</h3>
-        {fields.map((f, index) => (
-          <div key={f.id} className="border border-gray-200 p-4 rounded-md space-y-4">
-            <div className="grid grid-cols-3 gap-4">
-              <div>
-                <label className="block text-sm font-medium mb-1">Type</label>
-                <select className="w-full px-3 py-2 border border-gray-300 rounded-md" {...form.register(`elements.${index}.type` as const)}>
-                  <option value="text">Text</option>
-                  <option value="qrcode">QR Code</option>
-                  <option value="barcode">Barcode</option>
-                  <option value="image">Image</option>
-                </select>
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">X Position (mm)</label>
-                <Input type="number" step="0.01" {...form.register(`elements.${index}.x` as const, { valueAsNumber: true })} />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Y Position (mm)</label>
-                <Input type="number" step="0.01" {...form.register(`elements.${index}.y` as const, { valueAsNumber: true })} />
-              </div>
-            </div>
-            <div className="grid grid-cols-3 gap-4">
-              <div className="col-span-2">
-                <label className="block text-sm font-medium mb-1">Value / Placeholder / Image URL</label>
-                <Input placeholder="e.g., {equipment_name} or https://...img.png" {...form.register(`elements.${index}.value` as const)} />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Size</label>
-                <Input type="number" placeholder="Font/code/image width" {...form.register(`elements.${index}.size` as const, { valueAsNumber: true })} />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Height (image)</label>
-                <Input type="number" placeholder="Only for image" {...form.register(`elements.${index}.height` as const, { valueAsNumber: true })} />
-              </div>
-            </div>
-            <div className="flex items-end justify-between">
-              <div>
-                <label className="block text-sm font-medium mb-1">Color Override</label>
-                <Input type="color" {...form.register(`elements.${index}.color` as const)} />
-              </div>
-              <Button type="button" variant="destructive" onClick={() => remove(index)}>
-                Remove Element
-              </Button>
-            </div>
-          </div>
-        ))}
-        <Button type="button" variant="outline" onClick={() => append({ type: 'text', x: 0, y: 0, value: '', size: 12 })}>
-          Add Element
-        </Button>
-      </div>
-
-      {(() => {
-        const w = form.watch();
-        const previewTemplate: AssetTagTemplate = {
-          name: w.name,
-          description: w.description,
-          tagWidthMm: w.tagWidthMm,
-          tagHeightMm: w.tagHeightMm,
-          marginMm: w.marginMm as number | undefined,
-          backgroundColor: w.backgroundColor,
-          textColor: w.textColor,
-          borderColor: w.borderColor,
-          borderWidthMm: w.borderWidthMm as number | undefined,
-          textSizePt: w.textSizePt as number | undefined,
-          isMonochrome: w.isMonochrome,
-          prefix: w.prefix,
-          numberLength: w.numberLength as number | undefined,
-          suffix: w.suffix,
-          numberingScheme: w.numberingScheme,
-          stringTemplate: w.stringTemplate,
-          codeType: w.codeType,
-          codeSizeMm: w.codeSizeMm as number | undefined,
-          elements: ((w.elements || []) as unknown as FormValues['elements']).map((e) => ({
-            ...e,
-            size: e.size as number | undefined,
-            height: e.height as number | undefined,
-          })),
-        };
-        return (
-          <AssetTagTemplatePreview
-            template={previewTemplate}
-            editable
-            onElementsChange={(els: AssetTagTemplateElement[]) => form.setValue('elements', els as unknown as FormValues['elements'], { shouldDirty: true, shouldTouch: true })}
-          />
-        );
-      })()}
-
-      <Button type="submit" disabled={isLoading}>
-        {isLoading ? 'Updating...' : 'Update Template'}
-      </Button>
-    </form>
-  );
-}
